feat: Add updateJoueursEquipe method to EquipeCubit
The code changes include adding a new method `updateJoueursEquipe` to the `EquipeCubit` class. This method is responsible for updating the list of players in a team. It takes in the `joueursId`, `attente_joueursID`, and `equipeId` as parameters and makes an HTTP PUT request to the corresponding API endpoint. The response is then handled accordingly, either emitting a `UpdateJoueursEquipeStateGood` state or an `UpdateJoueursEquipeStateBad` state.

This commit message follows the established convention of starting with a verb in the imperative form, followed by a brief description of the changes made.